Restructured source files in mill.scalalib.bsp package
