+++ scalalib/src/mill/scalalib/bsp/JvmBuildTarget.scala
@@ -1,0 +1,10 @@
+package mill.scalalib.bsp
+
+case class JvmBuildTarget(
+    javaHome: Option[BspUri],
+    javaVersion: Option[String]
+)
+
+object JvmBuildTarget {
+  val dataKind: String = "jvm"
+}
